feat: add Agendamentos component and update routing :sparkles:
