--- src/components/HelloWorld.vue
@@ -1,17 +1,0 @@
-<script setup>
-import { ref } from "vue";
-
-defineProps({
-  msg: String,
-});
-
-const count = ref(0);
-</script>
-
-<template>nao fodo com nenhum de vcs</template>
-
-<style scoped>
-.read-the-docs {
-  color: #888;
-}
-</style>
